removed datastream, using kotlinx.serialization
