+++ api-gen/build.gradle
@@ -1,10 +1,13 @@
 apply plugin: 'kotlin'
+apply plugin: 'kotlinx-serialization'
 
 dependencies {
     compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
     compile "org.jetbrains.kotlin:kotlin-reflect:$kotlin_version"
     compile "com.stanfy:gson-xml-java:0.1.7"
     compile group: 'org.ogce', name: 'xpp3', version: '1.1.6'
+    compile "org.jetbrains.kotlinx:kotlinx-serialization-runtime:$serialization_version"
+
 }
 
 compileKotlin {
