+++ client/src/main/kotlin/builder/ui/BuildArea.kt
@@ -1,16 +1,16 @@
 package builder.ui
 
-import builder.Build
-import builder.data.Unit
+import ltd2.Build
+import ltd2.UnitState
 import kotlinx.html.id
 import react.RBuilder
 import react.dom.div
 
 interface BuildAreaEventHandler {
-    fun selectUnit(unit: Unit)
+    fun selectUnit(unit: UnitState)
 }
 
-fun RBuilder.buildArea(build: Build, selectedUnit: Unit?, eventHandler: BuildAreaEventHandler) {
+fun RBuilder.buildArea(build: Build, selectedUnit: UnitState?, eventHandler: BuildAreaEventHandler) {
     div {
         attrs.id = "build-area"
         unitList(build.getFighters().removeWorkersNotLevel(build.currentLevel), { true }, { eventHandler.selectUnit(it) }, selectedUnit)
