+++ client/src/main/kotlin/builder/ui/header/WaveInfo.kt
@@ -2,7 +2,7 @@ package builder.ui.header
 
 import app.format
 import ltd2.UnitDef
-import builder.data.Resistance
+import ltd2.Resistance
 import builder.ui.dpsUi
 import builder.ui.hpUi
 import kotlinx.html.js.onClickFunction
